Add "Open a new tab: First Child Tab" shortcut
This shortcut creates a new tab as the first child of the current tab.
As opposed to "Open a new tab: Child Tab" from TST, which starts the new
tab at the bottom of the parent tab's children, this command starts the
new tab at the top of the child list.